Move casting to IExpressionBuilder
